+++ contracts/protocol/Registry.sol
@@ -7,12 +7,9 @@ import "@openzeppelin/contracts-upgradeable/math/SafeMathUpgradeable.sol";
 import "@openzeppelin/contracts-upgradeable/utils/EnumerableSetUpgradeable.sol";
 import "@openzeppelin/contracts-upgradeable/utils/EnumerableMapUpgradeable.sol";
 
-
 import "contracts/interfaces/apwine/tokens/IAPWToken.sol";
 
-
-
-contract Registry is Initializable, AccessControlUpgradeable{
+contract Registry is Initializable, AccessControlUpgradeable {
     using SafeMathUpgradeable for uint256;
     using EnumerableSetUpgradeable for EnumerableSetUpgradeable.AddressSet;
     using EnumerableMapUpgradeable for EnumerableMapUpgradeable.UintToAddressMap;
@@ -34,20 +31,20 @@ contract Registry is Initializable, AccessControlUpgradeable{
     EnumerableSetUpgradeable.AddressSet private futureWalletsLogic;
     EnumerableSetUpgradeable.AddressSet private futuresLogic;
 
-    mapping(address=>string) private futureVaultName;
-    mapping(address=>string) private futureWalletName;
-    mapping(address=>string) private futureName;
+    mapping(address => string) private futureVaultName;
+    mapping(address => string) private futureWalletName;
+    mapping(address => string) private futureName;
 
     /* Futures Platforms Contracts */
     EnumerableSetUpgradeable.AddressSet private futurePlatformsDeployer;
-    mapping(address=>string) private futurePlatformsDeployerNames;
+    mapping(address => string) private futurePlatformsDeployerNames;
 
     string[] private futurePlatformsNames;
-    mapping(string=>address) private futurePlatformToDeployer;
-    mapping(string=>futurePlatform) private futurePlatformsName;
+    mapping(string => address) private futurePlatformToDeployer;
+    mapping(string => futurePlatform) private futurePlatformsName;
 
     /* Struct */
-    struct futurePlatform{
+    struct futurePlatform {
         address future;
         address futureVault;
         address futureWallet;
@@ -59,176 +56,251 @@ contract Registry is Initializable, AccessControlUpgradeable{
     address private APWineIBTLogic;
     address private FYTLogic;
 
-    event RegisteryUpdate(string _contractName,address _old, address _new);
-    event FuturePlatformAdded(address _futurePlatformDeployer, string _futurePlatformName, address _future, address _futureWallet, address _futureVault);
+    event RegisteryUpdate(string _contractName, address _old, address _new);
+    event FuturePlatformAdded(
+        address _futurePlatformDeployer,
+        string _futurePlatformName,
+        address _future,
+        address _futureWallet,
+        address _futureVault
+    );
 
-    
-    function initialize(address _admin) public initializer{
+    function initialize(address _admin) public initializer {
         _setupRole(DEFAULT_ADMIN_ROLE, _admin);
     }
 
     /* Setters */
-    function setTreasury(address _newTreasury) public{
+    function setTreasury(address _newTreasury) public {
         require(hasRole(ADMIN_ROLE, msg.sender), "Caller is not an admin");
-        emit RegisteryUpdate("Treasury", treasury,_newTreasury);
+        emit RegisteryUpdate("Treasury", treasury, _newTreasury);
         treasury = _newTreasury;
     }
 
-    function setGaugeController(address _newGaugeController) public{
+    function setGaugeController(address _newGaugeController) public {
         require(hasRole(ADMIN_ROLE, msg.sender), "Caller is not an admin");
-        emit RegisteryUpdate("GaugeController", gaugeController,_newGaugeController);
+        emit RegisteryUpdate(
+            "GaugeController",
+            gaugeController,
+            _newGaugeController
+        );
         gaugeController = _newGaugeController;
     }
 
-    function setController(address _newController) public{
+    function setController(address _newController) public {
         require(hasRole(ADMIN_ROLE, msg.sender), "Caller is not an admin");
-        emit RegisteryUpdate("Controller", controller,_newController);
+        emit RegisteryUpdate("Controller", controller, _newController);
         controller = _newController;
     }
 
-    function setAPW(address _newAPW) public{
+    function setAPW(address _newAPW) public {
         require(hasRole(ADMIN_ROLE, msg.sender), "Caller is not an admin");
-        emit RegisteryUpdate("APW", apw,_newAPW);
+        emit RegisteryUpdate("APW", apw, _newAPW);
         apw = _newAPW;
     }
 
     /* Getters */
-    function getDAOAddress() public view returns(address){
+    function getDAOAddress() public view returns (address) {
         return IAPWToken(apw).getDAO();
     }
 
-    function getAPWAddress() public view returns(address){
+    function getAPWAddress() public view returns (address) {
         return apw;
     }
 
-    function getVestingAddress() public view returns(address){
+    function getVestingAddress() public view returns (address) {
         return IAPWToken(apw).getVestingContract();
     }
 
-    function getControllerAddress() public view returns(address){
+    function getControllerAddress() public view returns (address) {
         return controller;
     }
 
-    function getTreasuryAddress() public view returns(address){
+    function getTreasuryAddress() public view returns (address) {
         return treasury;
     }
 
-    function getGaugeControllerAddress() public view returns(address){
+    function getGaugeControllerAddress() public view returns (address) {
         return gaugeController;
     }
 
     /* Logic setters*/
-    function setProxyFactory(address _proxyFactory) public{
+    function setProxyFactory(address _proxyFactory) public {
         require(hasRole(ADMIN_ROLE, msg.sender), "Caller is not an admin");
-        emit RegisteryUpdate("Proxy Factory", proxyFactory,_proxyFactory);
+        emit RegisteryUpdate("Proxy Factory", proxyFactory, _proxyFactory);
         proxyFactory = _proxyFactory;
     }
 
-    function setLiquidityGaugeLogic(address _liquidityGaugeLogic) public{
+    function setLiquidityGaugeLogic(address _liquidityGaugeLogic) public {
         require(hasRole(ADMIN_ROLE, msg.sender), "Caller is not an admin");
-        emit RegisteryUpdate("LiquidityGauge logic", liquidityGaugeLogic,_liquidityGaugeLogic);
+        emit RegisteryUpdate(
+            "LiquidityGauge logic",
+            liquidityGaugeLogic,
+            _liquidityGaugeLogic
+        );
         liquidityGaugeLogic = _liquidityGaugeLogic;
     }
 
-    function setAPWineIBTLogic(address _APWineIBTLogic) public{
+    function setAPWineIBTLogic(address _APWineIBTLogic) public {
         require(hasRole(ADMIN_ROLE, msg.sender), "Caller is not an admin");
-        emit RegisteryUpdate("APWineIBT logic", APWineIBTLogic,_APWineIBTLogic);
+        emit RegisteryUpdate(
+            "APWineIBT logic",
+            APWineIBTLogic,
+            _APWineIBTLogic
+        );
         APWineIBTLogic = _APWineIBTLogic;
     }
 
-    function setFYTLogic(address _FYTLogic) public{
+    function setFYTLogic(address _FYTLogic) public {
         require(hasRole(ADMIN_ROLE, msg.sender), "Caller is not an admin");
-        emit RegisteryUpdate("FYT  Logic", _FYTLogic,_FYTLogic);
+        emit RegisteryUpdate("FYT  Logic", _FYTLogic, _FYTLogic);
         FYTLogic = _FYTLogic;
     }
 
     /* Logic getters */
-    function getProxyFactoryAddress() public view returns(address){
+    function getProxyFactoryAddress() public view returns (address) {
         return proxyFactory;
     }
 
-    function getLiquidityGaugeLogicAddress() public view returns(address){
+    function getLiquidityGaugeLogicAddress() public view returns (address) {
         return liquidityGaugeLogic;
     }
 
-    function getAPWineIBTLogicAddress() public view returns(address){
+    function getAPWineIBTLogicAddress() public view returns (address) {
         return APWineIBTLogic;
     }
 
-    function getFYTLogicAddress() public view returns(address){
+    function getFYTLogicAddress() public view returns (address) {
         return FYTLogic;
     }
 
-
     /* Futures Deployer */
-    function addFuturePlatformDeployer(address _futurePlatformDeployer, string memory _futurePlatformDeployerName) public{
+    function addFuturePlatformDeployer(
+        address _futurePlatformDeployer,
+        string memory _futurePlatformDeployerName
+    ) public {
         require(hasRole(ADMIN_ROLE, msg.sender), "Caller is not an admin");
         futurePlatformsDeployer.add(_futurePlatformDeployer);
-        futurePlatformsDeployerNames[_futurePlatformDeployer]=_futurePlatformDeployerName;
+        futurePlatformsDeployerNames[
+            _futurePlatformDeployer
+        ] = _futurePlatformDeployerName;
     }
 
-    function isRegisteredFuturePlatformDeployer(address _futurePlatformDeployer) public view returns(bool){
+    function isRegisteredFuturePlatformDeployer(address _futurePlatformDeployer)
+        public
+        view
+        returns (bool)
+    {
         return futurePlatformsDeployer.contains(_futurePlatformDeployer);
     }
-    
-    function getFuturePlatformDeployerAt(uint256 _index) external view returns(address){
+
+    function getFuturePlatformDeployerAt(uint256 _index)
+        external
+        view
+        returns (address)
+    {
         return futurePlatformsDeployer.at(_index);
     }
 
-    function futurePlatformDeployerCount() external view returns(uint256) {
+    function futurePlatformDeployerCount() external view returns (uint256) {
         return futurePlatformsDeployer.length();
     }
 
-    function getFuturePlatformDeployerName(address _futurePlatformDeployer) external view returns(string memory){
-        require(futurePlatformsDeployer.contains(_futurePlatformDeployer), "invalid future platform deployer");
+    function getFuturePlatformDeployerName(address _futurePlatformDeployer)
+        external
+        view
+        returns (string memory)
+    {
+        require(
+            futurePlatformsDeployer.contains(_futurePlatformDeployer),
+            "invalid future platform deployer"
+        );
         return futurePlatformsDeployerNames[_futurePlatformDeployer];
     }
 
-
     /* Future Platform */
-    function addFuturePlatform(address _futurePlatformDeployer, string memory _futurePlatformName, address _future, address _futureWallet, address _futureVault) public{
+    function addFuturePlatform(
+        address _futurePlatformDeployer,
+        string memory _futurePlatformName,
+        address _future,
+        address _futureWallet,
+        address _futureVault
+    ) public {
         require(hasRole(ADMIN_ROLE, msg.sender), "Caller is not an admin");
-        require(futurePlatformsDeployer.contains(_futurePlatformDeployer), "invalid future platfrom deployer address");
+        require(
+            futurePlatformsDeployer.contains(_futurePlatformDeployer),
+            "invalid future platfrom deployer address"
+        );
 
-        futurePlatform memory newFuturePlaform = futurePlatform({
-            futureVault:_futureVault,
-            futureWallet:_futureWallet,
-            future:_future
-        });
-        
+        futurePlatform memory newFuturePlaform =
+            futurePlatform({
+                futureVault: _futureVault,
+                futureWallet: _futureWallet,
+                future: _future
+            });
 
-        if(!isRegisteredFuturePlatform(_futurePlatformName)) futurePlatformsNames.push(_futurePlatformName);
-       
+        if (!isRegisteredFuturePlatform(_futurePlatformName))
+            futurePlatformsNames.push(_futurePlatformName);
+
         futurePlatformsName[_futurePlatformName] = newFuturePlaform;
         futurePlatformToDeployer[_futurePlatformName] = _futurePlatformDeployer;
-        emit FuturePlatformAdded(_futurePlatformDeployer, _futurePlatformName, _future, _futureWallet,_futureVault);    
+        emit FuturePlatformAdded(
+            _futurePlatformDeployer,
+            _futurePlatformName,
+            _future,
+            _futureWallet,
+            _futureVault
+        );
     }
 
-    function removeFuturePlatform(string memory _futurePlatformName) public{
+    function removeFuturePlatform(string memory _futurePlatformName) public {
         require(hasRole(ADMIN_ROLE, msg.sender), "Caller is not an admin");
-        require(isRegisteredFuturePlatform(_futurePlatformName), "invalid future platform name");
-        
-        for(uint i=0;i<futurePlatformsNames.length;i++){ // can be optimized
-            if(keccak256(bytes(futurePlatformsNames[i]))==keccak256(bytes(_futurePlatformName))){
+        require(
+            isRegisteredFuturePlatform(_futurePlatformName),
+            "invalid future platform name"
+        );
+
+        for (uint256 i = 0; i < futurePlatformsNames.length; i++) {
+            // can be optimized
+            if (
+                keccak256(bytes(futurePlatformsNames[i])) ==
+                keccak256(bytes(_futurePlatformName))
+            ) {
                 delete futurePlatformsNames[i];
                 break;
             }
         }
 
         delete futurePlatformToDeployer[_futurePlatformName];
-        delete futurePlatformsName[_futurePlatformName];      
+        delete futurePlatformsName[_futurePlatformName];
     }
 
-    function isRegisteredFuturePlatform(string memory _futurePlatformName) public view returns(bool){
-        for(uint i=0;i<futurePlatformsNames.length;i++){
-            if(keccak256(bytes(futurePlatformsNames[i]))==keccak256(bytes(_futurePlatformName))) return true;
+    function isRegisteredFuturePlatform(string memory _futurePlatformName)
+        public
+        view
+        returns (bool)
+    {
+        for (uint256 i = 0; i < futurePlatformsNames.length; i++) {
+            if (
+                keccak256(bytes(futurePlatformsNames[i])) ==
+                keccak256(bytes(_futurePlatformName))
+            ) return true;
         }
         return false;
     }
-    
-    function getFuturePlatform(string memory _futurePlatformName) public view returns(address[3] memory){
-        futurePlatform  memory futurePlatformContracts = futurePlatformsName[_futurePlatformName];
-        address[3] memory addressesArrays = [futurePlatformContracts.future, futurePlatformContracts.futureVault, futurePlatformContracts.futureWallet];
+
+    function getFuturePlatform(string memory _futurePlatformName)
+        public
+        view
+        returns (address[3] memory)
+    {
+        futurePlatform memory futurePlatformContracts =
+            futurePlatformsName[_futurePlatformName];
+        address[3] memory addressesArrays =
+            [
+                futurePlatformContracts.future,
+                futurePlatformContracts.futureVault,
+                futurePlatformContracts.futureWallet
+            ];
         return addressesArrays;
     }
 
@@ -236,26 +308,26 @@ contract Registry is Initializable, AccessControlUpgradeable{
         return futurePlatformsNames.length;
     }
 
-    function getFuturePlatformNames() external view returns(string[] memory){
+    function getFuturePlatformNames() external view returns (string[] memory) {
         return futurePlatformsNames;
     }
 
     /* Futures */
-    function addFuture(address _future) public returns(bool){
+    function addFuture(address _future) public returns (bool) {
         require(hasRole(CONTROLLER_ROLE, msg.sender), "Caller is not an admin");
         return futures.add(_future);
     }
-    
-    function removeFuture(address _future) public returns(bool){
+
+    function removeFuture(address _future) public returns (bool) {
         require(hasRole(CONTROLLER_ROLE, msg.sender), "Caller is not an admin");
         return futures.remove(_future);
     }
 
-    function isRegisteredFuture(address _future) external view returns(bool){
-        return futures.contains(_future); 
+    function isRegisteredFuture(address _future) external view returns (bool) {
+        return futures.contains(_future);
     }
 
-    function getFutureAt(uint256 _index) external view returns(address){
+    function getFutureAt(uint256 _index) external view returns (address) {
         return futures.at(_index);
     }
 
@@ -263,13 +335,16 @@ contract Registry is Initializable, AccessControlUpgradeable{
         return futures.length();
     }
 
-
     /* Future Wallets Logic */
-    function isRegisteredFutureWallet(address _futureWallet) external view returns(bool){
-        return futureWalletsLogic.contains(_futureWallet); 
+    function isRegisteredFutureWallet(address _futureWallet)
+        external
+        view
+        returns (bool)
+    {
+        return futureWalletsLogic.contains(_futureWallet);
     }
 
-    function getFuturWalletAt(uint256 _index) external view returns(address){
+    function getFuturWalletAt(uint256 _index) external view returns (address) {
         return futureWalletsLogic.at(_index);
     }
 
@@ -277,17 +352,28 @@ contract Registry is Initializable, AccessControlUpgradeable{
         return futureWalletsLogic.length();
     }
 
-    function getFutureWalletName(address _futureWalletAddress) external view returns(string memory){
-        require(futureWalletsLogic.contains(_futureWalletAddress),"invalid address");
+    function getFutureWalletName(address _futureWalletAddress)
+        external
+        view
+        returns (string memory)
+    {
+        require(
+            futureWalletsLogic.contains(_futureWalletAddress),
+            "invalid address"
+        );
         return futureWalletName[_futureWalletAddress];
     }
 
     /* Future Vault Logic*/
-    function isRegisteredFutureVault(address _futureVault) public view returns(bool){
-        return futureWalletsLogic.contains(_futureVault); 
+    function isRegisteredFutureVault(address _futureVault)
+        public
+        view
+        returns (bool)
+    {
+        return futureWalletsLogic.contains(_futureVault);
     }
 
-    function getFutureVaultAt(uint256 _index) public view returns(address){
+    function getFutureVaultAt(uint256 _index) public view returns (address) {
         return futureWalletsLogic.at(_index);
     }
 
@@ -295,17 +381,25 @@ contract Registry is Initializable, AccessControlUpgradeable{
         return futureWalletsLogic.length();
     }
 
-    function getFutureVaultName(address _futureVault) external view returns(string memory){
-        require(futureWalletsLogic.contains(_futureVault),"invalid address");
+    function getFutureVaultName(address _futureVault)
+        external
+        view
+        returns (string memory)
+    {
+        require(futureWalletsLogic.contains(_futureVault), "invalid address");
         return futureWalletName[_futureVault];
     }
 
     /* Future Logic*/
-    function isRegisteredFutureLogic(address _futureLogic) external view returns(bool){
-        return futuresLogic.contains(_futureLogic); 
+    function isRegisteredFutureLogic(address _futureLogic)
+        external
+        view
+        returns (bool)
+    {
+        return futuresLogic.contains(_futureLogic);
     }
 
-    function getFutureLogicAt(uint256 _index) external view returns(address){
+    function getFutureLogicAt(uint256 _index) external view returns (address) {
         return futuresLogic.at(_index);
     }
 
@@ -313,10 +407,12 @@ contract Registry is Initializable, AccessControlUpgradeable{
         return futuresLogic.length();
     }
 
-    function getFutureLogicName(address _futureLogicAddress) external view returns(string memory){
-        require(futuresLogic.contains(_futureLogicAddress),"invalid address");
+    function getFutureLogicName(address _futureLogicAddress)
+        external
+        view
+        returns (string memory)
+    {
+        require(futuresLogic.contains(_futureLogicAddress), "invalid address");
         return futureWalletName[_futureLogicAddress];
     }
-
-
 }
